+++ src/corelibs/U2Core/src/tasks/SequenceWalkerTask.cpp
@@ -37,7 +37,7 @@ SequenceWalkerConfig::SequenceWalkerConfig()
 SequenceWalkerTask::SequenceWalkerTask(const SequenceWalkerConfig& c, SequenceWalkerCallback* cb, const QString& name, TaskFlags tf) 
 : Task(name, tf), config(c), callback(cb)
 {
-    assert(config.chunkSize > config.overlapSize); // if chunk == overlap -> infinite loop occurs
+    assert(config.chunkSize > static_cast<uint>(config.overlapSize)); // if chunk == overlap -> infinite loop occurs
     assert(cb != NULL);
     assert(config.strandToWalk == StrandOption_DirectOnly || config.complTrans!=NULL);
